scripts/requirements.txt
openai[datalib]
azure-identity
python-dotenv